+++ spring/src/main/java/com/playground/repository/HoursRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HoursRepository extends JpaRepository<Hours, Long> {
+public interface HoursRepository extends JpaRepository<Hours, Integer> {
 
 }
