+++ spring/src/main/java/com/playground/repository/PlaygroundRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlaygroundRepository extends JpaRepository<Playground, Long> {
+public interface PlaygroundRepository extends JpaRepository<Playground, Integer> {
 
 }
